rename files and refactor api code into service classes; implement refresh token 50%
