+++ src/pages/HomeLayoutComponent.scss
